Finished refactering most of the code Coverage is expected to work this time [Finishes Delivers #162306468]
