+++ .coverage
@@ -1,0 +1,1 @@
+!coverage.py: This is a private format, don't read it directly!{"arcs":{"/home/paphra/andela/iReporter/api/v1/app.py":[[86,87],[92,93],[40,41],[18,19],[202,203],[-151,152],[79,78],[96,97],[43,42],[6,7],[-137,138],[172,173],[51,52],[25,26],[28,29],[194,-184],[35,36],[-48,49],[17,18],[213,218],[108,110],[29,30],[79,80],[10,13],[113,-107],[141,151],[39,48],[145,148],[204,-200],[69,-67],[16,17],[33,34],[41,45],[184,200],[68,70],[127,137],[173,174],[111,112],[153,152],[225,-223],[218,223],[112,-107],[110,111],[80,-77],[15,16],[109,-107],[203,204],[22,23],[191,192],[84,107],[195,-184],[71,72],[111,110],[160,168],[171,179],[193,194],[94,-84],[151,158],[58,59],[-141,143],[20,21],[95,96],[201,202],[122,-116],[165,166],[60,61],[71,74],[-39,40],[200,208],[103,104],[-84,86],[100,101],[144,145],[72,73],[67,77],[-2,2],[158,171],[64,-55],[101,102],[72,71],[23,24],[-67,68],[179,184],[81,-77],[90,91],[205,-200],[131,132],[-184,185],[57,58],[55,67],[20,28],[132,131],[190,191],[155,-151],[59,60],[-55,57],[116,127],[129,131],[42,43],[168,-158],[189,190],[61,62],[45,-39],[62,-55],[-208,210],[-77,78],[-13,15],[174,175],[172,176],[48,51],[26,-13],[74,-67],[32,33],[120,121],[147,148],[152,155],[220,-218],[153,154],[185,186],[148,-141],[138,-137],[77,84],[68,69],[-116,118],[202,201],[192,193],[97,99],[88,103],[137,141],[132,133],[41,42],[88,89],[102,-84],[144,147],[119,120],[128,129],[52,55],[188,189],[163,164],[154,-151],[-223,225],[19,20],[-158,160],[107,116],[30,31],[-218,220],[43,44],[49,-48],[143,144],[2,4],[-200,201],[78,81],[118,119],[5,6],[161,162],[166,-158],[208,213],[98,99],[59,63],[133,-127],[-107,108],[63,64],[-171,172],[88,95],[24,25],[173,172],[34,-13],[36,-13],[104,-84],[108,109],[31,32],[44,-39],[188,195],[89,90],[4,5],[160,161],[131,134],[91,92],[228,-2],[73,71],[110,113],[93,94],[42,45],[175,-171],[162,163],[9,10],[210,-208],[13,39],[78,79],[164,165],[70,71],[-127,128],[87,88],[186,187],[7,9],[97,98],[201,205],[121,122],[21,22],[187,188],[134,-127],[99,100],[152,153],[223,228],[17,35],[176,-171],[-179,181],[181,-179]],"/home/paphra/andela/iReporter/api/v1/__init__.py":[[-1,1],[1,-1]],"/home/paphra/andela/iReporter/api/tests/test_api1.py":[[216,222],[54,55],[-242,243],[92,93],[276,277],[256,262],[40,41],[273,-267],[94,95],[174,-160],[287,-286],[42,43],[-267,268],[127,128],[260,-256],[96,97],[56,57],[234,235],[217,218],[122,123],[164,165],[44,45],[43,44],[210,211],[250,251],[213,214],[211,212],[-216,217],[25,26],[36,37],[59,60],[82,83],[240,-237],[27,28],[171,172],[75,77],[33,34],[68,69],[7,10],[166,167],[29,30],[173,174],[177,178],[105,106],[228,237],[142,143],[97,98],[231,232],[284,-275],[195,196],[268,269],[55,56],[17,18],[72,73],[109,110],[91,92],[265,-247],[146,147],[112,113],[140,141],[10,289],[158,-12],[184,185],[18,19],[52,53],[188,204],[67,68],[205,206],[218,219],[-188,189],[106,107],[201,-188],[126,127],[69,70],[107,108],[196,197],[-237,238],[243,244],[252,253],[153,155],[60,61],[277,278],[258,259],[71,72],[-247,248],[-10,10],[-275,276],[65,66],[58,59],[156,158],[172,173],[20,21],[117,118],[280,284],[286,-10],[165,166],[101,102],[66,67],[248,249],[103,104],[144,146],[100,101],[-204,205],[121,122],[21,22],[238,239],[139,140],[37,38],[138,139],[-2,2],[73,74],[50,51],[46,47],[167,168],[134,135],[83,84],[245,-242],[3,5],[-160,161],[113,114],[90,91],[269,270],[259,260],[77,78],[131,132],[-209,210],[5,7],[194,195],[253,256],[49,50],[57,58],[84,85],[237,242],[229,230],[80,81],[257,258],[62,63],[289,-2],[34,35],[155,156],[74,75],[189,190],[28,29],[219,220],[163,164],[233,234],[30,31],[13,14],[212,213],[150,151],[242,247],[220,-216],[95,96],[267,275],[61,62],[16,17],[119,120],[197,-193],[168,169],[45,46],[207,-204],[23,24],[170,171],[120,121],[185,186],[263,265],[39,40],[230,231],[87,89],[-176,177],[81,82],[132,133],[279,280],[193,199],[41,42],[271,272],[160,176],[2,3],[-193,194],[102,103],[51,52],[128,129],[136,137],[190,191],[14,16],[-256,257],[270,271],[53,54],[35,36],[110,111],[85,86],[226,-209],[244,245],[79,80],[130,131],[-228,229],[118,119],[222,224],[12,160],[133,134],[247,267],[178,179],[161,162],[176,183],[104,105],[116,117],[98,99],[143,144],[-183,184],[31,32],[251,252],[214,216],[129,130],[24,25],[278,279],[70,71],[-286,287],[63,64],[239,240],[10,12],[186,-183],[89,90],[124,125],[204,209],[125,126],[47,48],[262,263],[99,100],[224,226],[209,228],[32,33],[115,116],[-12,13],[111,112],[147,148],[64,65],[93,94],[249,250],[162,163],[137,138],[22,23],[78,79],[38,39],[235,-228],[19,20],[86,87],[199,201],[48,49],[169,170],[108,109],[26,27],[141,142],[206,207],[149,150],[135,136],[151,152],[152,153],[148,149],[183,188],[179,-176],[114,115],[232,233],[191,193],[123,124],[275,286],[272,273]],"/home/paphra/andela/iReporter/api/tests/__init__.py":[[-1,1],[1,-1]],"/home/paphra/andela/iReporter/api/tests/test_api2.py":[[60,-59],[-40,41],[10,-2],[52,57],[41,-40],[24,26],[-34,35],[18,19],[-37,38],[-43,44],[-10,10],[-31,32],[49,50],[50,51],[20,21],[23,24],[10,12],[35,-34],[37,40],[7,10],[57,-43],[29,-28],[15,17],[47,48],[-12,14],[43,59],[-28,29],[32,-31],[28,31],[21,23],[26,-12],[40,43],[-2,2],[14,15],[31,34],[38,-37],[34,37],[19,20],[59,-10],[3,5],[48,49],[12,28],[45,46],[17,18],[5,7],[44,45],[2,3],[46,47],[-59,60],[51,52]],"/home/paphra/andela/iReporter/api/__init__.py":[[-1,1],[1,-1]]}}
