+++ requirements.txt
@@ -1,76 +1,11 @@
-astroid==2.0.4
-atomicwrites==1.2.1
-attrs==18.2.0
-backcall==0.1.0
-bcrypt==3.1.4
-certifi==2018.4.16
-cffi==1.11.5
-chardet==3.0.4
-click==6.7
-colorama==0.3.9
-coverage==4.5.1
+Click==7.0
+coverage==4.5.2
 coveralls==1.5.0
-cycler==0.10.0
-decorator==4.3.0
-docopt==0.6.2
-flake8==3.5.0
 Flask==1.0.2
-Flask-Bcrypt==0.7.1
-Flask-GoogleMaps==0.2.6
-Flask-Testing==0.7.1
-Flask-WTF==0.14.2
 future==0.16.0
-gunicorn==19.9.0
-idna==2.7
-ipython==6.4.0
-ipython-genutils==0.2.0
-isort==4.3.4
 itsdangerous==0.24
-jedi==0.12.1
 Jinja2==2.10
-kiwisolver==1.0.1
-lazy-object-proxy==1.3.1
 MarkupSafe==1.0
-matplotlib==2.2.2
-mccabe==0.6.1
-more-itertools==4.3.0
-numpy==1.14.4
-oauthlib==2.1.0
-parso==0.3.1
-pathlib2==2.3.2
-patsy==0.5.0
-pickleshare==0.7.4
-pluggy==0.7.1
-prompt-toolkit==1.0.15
-psycopg2==2.7.5
-py==1.6.0
-pycodestyle==2.3.1
-pycparser==2.19
-pycurl==7.43.0
-pyflakes==1.6.0
-Pygments==2.2.0
-PyJWT==1.6.4
-pylint==2.1.1
-pyparsing==2.2.0
-PySocks==1.6.8
 pytest==4.0.0
 pytest-cov==2.6.0
-python-dateutil==2.7.3
-pytz==2018.4
-requests==2.19.1
-requests-oauthlib==1.0.0
-scandir==1.9.0
-scipy==1.1.0
-simplegeneric==0.8.1
-six==1.11.0
-statsmodels==0.9.0
-tinydb==3.9.0.post1
-traitlets==4.3.2
-tuspy==0.2.1
-typed-ast==1.1.0
-typing==3.6.4
-virtualenv==16.0.0
-wcwidth==0.1.7
 Werkzeug==0.14.1
-win-unicode-console==0.5
-wrapt==1.10.11
